[7.2]GFAlertVC, Convenience Init, DateDecoding, Networking Refactor
Cleaning up our GFAlertVC
Convenience Initializers
DateDecodingStrategy
Refactoring GFAvatarImageView
GFDataLoadingVC

ISO 8601:

https://en.wikipedia.org/wiki/ISO_8601

More reading on Convenience Initializers:

https://useyourloaf.com/blog/adding-swift-convenience-initializers/
