+++ gitHubFollowers/Custom Views/ImageViews/GFAvatarImageView.swift
@@ -1,0 +1,31 @@
+//
+//  GFAvatarImageView.swift
+//  gitHubFollowers
+//
+//  Created by Lucian Cerbu on 27.12.2021.
+//
+
+import UIKit
+
+class GFAvatarImageView: UIImageView {
+    
+    let cache = NetworkManager.shared.cache
+    let placeholderImage = UIImage(named: "avatar-placeholder")!
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        configure()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configure() {
+        layer.cornerRadius = 10
+        clipsToBounds = true
+        image = placeholderImage
+        translatesAutoresizingMaskIntoConstraints = false
+    }
+}
