+++ gitHubFollowers/Custom Views/Labels/GFSecondaryLabel.swift
@@ -11,6 +11,7 @@ class GFSecondaryLabel: UILabel {
 
     override init(frame: CGRect) {
         super.init(frame: frame)
+        configure()
     }
     
     
@@ -19,11 +20,9 @@ class GFSecondaryLabel: UILabel {
     }
     
     
-    init(fontSize: CGFloat) {
-        super.init(frame: .zero)
+    convenience init(fontSize: CGFloat) {
+        self.init(frame: .zero)
         font = UIFont.systemFont(ofSize: fontSize, weight: .medium)
-        
-        configure()
     }
