+++ gitHubFollowers/Screens/FavoritesListVC.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class FavoritesListVC: UIViewController {
+class FavoritesListVC: GFDataLoadingVC {
 
     let tableView = UITableView()
     var favorites: [Follower] = []
